requirements.txt
python-dotenv
groq
tavily-python
google-generativeai
newsapi-python
duckduckgo-search


pinecone       
streamlit
sentence-transformers